+++ backend/src/main/java/ca/metricalsky/winston/events/FetchDataEventBuilder.java
@@ -1,0 +1,11 @@
+package ca.metricalsky.winston.events;
+
+import ca.metricalsky.winston.service.fetch.FetchResult;
+
+public class FetchDataEventBuilder implements EventBuilder {
+
+    @Override
+    public Object buildEvent(FetchResult fetchResult) {
+        return FetchDataEvent.of(fetchResult);
+    }
+}