+++ backend/src/main/java/ca/metricalsky/winston/events/SsePublisher.java
@@ -1,7 +1,9 @@
 package ca.metricalsky.winston.events;
 
+import ca.metricalsky.winston.service.fetch.FetchResult;
 import com.google.common.annotations.VisibleForTesting;
 import lombok.Getter;
+import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.http.ProblemDetail;
@@ -17,6 +19,9 @@ public class SsePublisher {
     private final UUID id;
     private final SseEmitter sseEmitter;
     private boolean open;
+
+    @Setter
+    private EventBuilder eventBuilder;
 
     public SsePublisher(Long timeout) {
         this(new SseEmitter(timeout));
@@ -46,7 +51,11 @@ public class SsePublisher {
                 .data(subscriptionEvent, MediaType.APPLICATION_JSON));
     }
 
-    public void publish(FetchDataEvent fetchDataEvent) {
+    public void publish(FetchResult fetchResult) {
+        publish(eventBuilder.buildEvent(fetchResult));
+    }
+
+    public void publish(Object fetchDataEvent) {
         publish(SseEmitter.event()
                 .id(UUID.randomUUID().toString())
                 .name("fetch-data")
@@ -67,7 +76,7 @@ public class SsePublisher {
                 .data(error, MediaType.APPLICATION_PROBLEM_JSON));
     }
 
-    private void publish(SseEmitter.SseEventBuilder builder) {
+    protected void publish(SseEmitter.SseEventBuilder builder) {
         if (open) {
             try {
                 sseEmitter.send(builder);